+++ plugin-src/translators/text/index.ts
@@ -1,1 +1,2 @@
+export * from './buildTextContent';
 export * from './translateTextSegments';
