+++ plugin-src/translators/vectors/index.ts
@@ -1,6 +1,4 @@
-export * from './applyMatrixToCommand';
 export * from './normalizeCommands';
+export * from './serializeCommands';
 export * from './translateCommands';
-export * from './translateNonRotatedCommands';
-export * from './translateRotatedCommands';
 export * from './translateWindingRule';
