+++ plugin-src/utils/index.ts
@@ -1,9 +1,11 @@
+export * from './applyMatrixToCommand';
 export * from './applyMatrixToPoint';
 export * from './applyRotation';
 export * from './calculateLinearGradient';
 export * from './calculateRadialGradient';
 export * from './editorType';
 export * from './generateUuid';
-export * from './progress';
 export * from './matrixInvert';
+export * from './multiplyMatrix';
+export * from './progress';
 export * from './rgbToHex';
